code

+++ database/migrations/2022_09_18_141044_order.php
@@ -21,11 +21,11 @@ return new class extends Migration
             $table->string('phone');
             $table->string('email');
             $table->bigInteger('id_coupons')->default(0);
-            $table->integer('coupons_price')->default(0);
-            $table->integer('total_price')->default(0);
-            $table->integer('ship')->default(0);
-            $table->integer('id_coupon_frs')->default(0);
-            $table->integer('coupons_frs_price')->default(0);
+            $table->bigInteger('coupons_price')->default(0);
+            $table->bigInteger('total_price')->default(0);
+            $table->bigInteger('ship')->default(0);
+            $table->bigInteger('id_coupon_frs')->default(0);
+            $table->bigInteger('coupons_frs_price')->default(0);
             $table->enum('payment_method',[0,1,2,3])->comment('0 là tt khi nhận, 1 là qua user pay, 2 là qua vnpay, 3 là qua momo')->default(0);
             $table->enum('payment_status', [0,1])->comment('0 là chưa thanh toán, 1 là đã thanh toán');
             $table->enum('status_order',[0,1,2,3,4])->comment('0 là chờ xử lí, 1 là đã xác nhận - đang xử lí, 2 là đang giao hàng, 3 là thành công, 4 là thất bại')->default(0);
